pyproject.toml
[tool.poetry]
name = "alice-yamaha-skill"
version = "0.1.0"
description = "Yamaha receiver skill for Alice Yandex"
authors = ["Anton Bazhanov <anton.bazhanov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.45.0"
rxv = "^0.6.0"
uvicorn = "^0.11.1"
httpx = "^0.10.1"
pyaml = "^19.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
